📦 New: middleware and cookie

+++ pages/api/console/users/remove_allowlist.ts
@@ -1,17 +1,18 @@
 import { ok } from 'lib/response';
 import prisma from "lib/prisma"
+import checkAuth from '../middlerware/checkAuth';
 
-export default async (req, res) => {
-    // middleware validate user authentication
+export default checkAuth(async (req, res) => {
 
     const { id } = req.body;
 
-    await prisma.allowlist.delete({
-        where: {
-            allowlist_id: id
-        }
-    })
+    if (id)
+        await prisma.allowlist.delete({
+            where: {
+                allowlist_id: id
+            }
+        })
 
     return ok(res);
 
-}
+})
