+++ pages/api/console/users/remove_blocklist.ts
@@ -1,17 +1,19 @@
 import { ok } from 'lib/response';
 import prisma from "lib/prisma"
+import checkAuth from '../middlerware/checkAuth';
 
-export default async (req, res) => {
-    // middleware validate user authentication
+export default checkAuth(async (req, res) => {
+
 
     const { id } = req.body;
 
-    await prisma.blocklist.delete({
-        where: {
-            blocklist_id: id
-        }
-    })
+    if (id)
+        await prisma.blocklist.delete({
+            where: {
+                blocklist_id: id
+            }
+        })
 
     return ok(res);
 
-}
+})
